perf: improve code comprehension
feat: categories.title is not null, expenses.amount is decimal

refactor: update schemas

test: update tests

refactor: remove redundant imports

refactor: replace filter with filter_by

refactor: add type annotation

+++ alembic/versions/37080222be44_add_expenses_table.py
@@ -22,7 +22,7 @@ def upgrade() -> None:
         "expenses",
         sa.Column("id", sa.Integer(), nullable=False),
         sa.Column("descriptions", sa.String(), nullable=False),
-        sa.Column("amount", sa.Float(), nullable=False),
+        sa.Column("amount", sa.DECIMAL(), nullable=False),
         sa.Column("time", sa.DateTime(), nullable=True),
         sa.Column("user_id", sa.Integer(), nullable=False),
         sa.Column("group_id", sa.Integer(), nullable=False),
