+++ src/models/expense.py
@@ -5,7 +5,7 @@ from sqlalchemy import (
     Integer,
     String,
     ForeignKey,
-    Float,
+    DECIMAL,
     DateTime,
     ForeignKeyConstraint,
 )
@@ -20,7 +20,7 @@ class Expense(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     descriptions = Column(String, nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(DECIMAL, nullable=False)
     time = Column(DateTime, default=datetime.datetime.utcnow(), nullable=False)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     group_id = Column(Integer, index=True, nullable=False)
